Move SendMsg/RecvMsg traits and impls to msg submodule.
