+++ audioipc/src/lib.rs
@@ -40,11 +40,8 @@ pub mod messages;
 mod msg;
 pub mod shm;
 
-use iovec::IoVec;
 pub use messages::{ClientMessage, ServerMessage};
 use std::env::temp_dir;
-use std::io;
-use std::os::unix::io::AsRawFd;
 use std::path::PathBuf;
 
 // This must match the definition of
@@ -53,41 +50,6 @@ use std::path::PathBuf;
 pub type PlatformHandleType = *mut std::os::raw::c_void;
 #[cfg(not(target_os = "windows"))]
 pub type PlatformHandleType = libc::c_int;
-
-// Extend sys::os::unix::net::UnixStream to support sending and receiving a single file desc.
-// We can extend UnixStream by using traits, eliminating the need to introduce a new wrapped
-// UnixStream type.
-trait RecvMsg {
-    fn recv_msg(
-        &mut self,
-        iov: &mut [&mut IoVec],
-        cmsg: &mut [u8],
-    ) -> io::Result<(usize, usize, i32)>;
-}
-
-trait SendMsg {
-    fn send_msg(&mut self, iov: &[&IoVec], cmsg: &[u8]) -> io::Result<usize>;
-}
-
-impl<T: AsRawFd> RecvMsg for T {
-    fn recv_msg(
-        &mut self,
-        iov: &mut [&mut IoVec],
-        cmsg: &mut [u8],
-    ) -> io::Result<(usize, usize, i32)> {
-        #[cfg(target_os = "linux")]
-        let flags = libc::MSG_CMSG_CLOEXEC;
-        #[cfg(not(target_os = "linux"))]
-        let flags = 0;
-        msg::recv_msg_with_flags(self.as_raw_fd(), iov, cmsg, flags)
-    }
-}
-
-impl<T: AsRawFd> SendMsg for T {
-    fn send_msg(&mut self, iov: &[&IoVec], cmsg: &[u8]) -> io::Result<usize> {
-        msg::send_msg_with_flags(self.as_raw_fd(), iov, cmsg, 0)
-    }
-}
 
 pub fn get_shm_path(dir: &str) -> PathBuf {
     let pid = unsafe { libc::getpid() };
